Avoid infinite loop when reading an empty ECU version file
The current algorithm reads data in a while loop that stops only when
the end-of-file is reached. Each call to `fread` tries to read the exact
file length.

This works well, except when the file length is 0. In this precise
situation, we keep reading 0 bytes from the file, thusly never reaching
the end of the file.

This can be tested with the following extremely reduced example:

```c
// Filepath: main.c

int main(int argc, char *argv[]) {
   assert(argc == 2);

   char *path = argv[1];
   FILE *f = fopen(path, "r");
   assert(f != NULL);
   int fd = fileno(f);

   struct stat st;
   int rc = fstat(fd, &st);
   assert(rc == 0);

   size_t size = st.st_size;

   // We don't actually care about the buffer content - we just need a
   // valid buffer to pass to `fread`.
   char *buffer = malloc(size);
   assert(buffer != NULL);

   while (!feof(f)) {
     fread(buffer, 1, size, f);
     assert(!ferror(f));
   }

   fclose(f);
}
```

It can be tested with the following:

```
$ gcc main.c -o main
$ rm empty.txt ; touch empty.txt
$ ./main empty.txt
<program never ends>
```

This commit fixes this by reading one more byte at each iteration. That
way we are guaranteed to read at least one byte per iteration, thus
making progress and never hanging in this loop.

Signed-off-by: Sasha Pourcelot <sasha.pourcelot@protonmail.com>